+++ src/components/Proposal/Calls/RecommendedCalls.tsx
@@ -1,0 +1,85 @@
+import { normalizeBalance } from 'utils';
+import { BlockchainLink } from 'components/common';
+export const RecommendedCalls = ({
+  to,
+  from,
+  recommendedCallUsed,
+  callParameters,
+  encodedFunctionName,
+  data,
+  // value,
+  // contractABI,
+  showMore,
+}) => {
+  let decodedCallText = '';
+
+  if (
+    recommendedCallUsed.decodeText &&
+    recommendedCallUsed.decodeText.length > 0
+  ) {
+    decodedCallText = recommendedCallUsed.decodeText;
+
+    recommendedCallUsed.params.map((_, paramIndex) => {
+      if (recommendedCallUsed.params[paramIndex].decimals) {
+        decodedCallText = decodedCallText.replaceAll(
+          '[PARAM_' + paramIndex + ']',
+          String(
+            normalizeBalance(
+              callParameters[paramIndex],
+              recommendedCallUsed.params[paramIndex].decimals
+            )
+          )
+        );
+      }
+      decodedCallText = decodedCallText.replaceAll(
+        '[PARAM_' + paramIndex + ']',
+        callParameters[paramIndex]
+      );
+    });
+  }
+  if (showMore) {
+    return (
+      <div>
+        <p>
+          <strong>From: </strong>{' '}
+          <small>
+            <BlockchainLink text={from} toCopy={false} />
+          </small>
+        </p>
+        <p>
+          <strong>To: </strong>{' '}
+          <small>
+            <BlockchainLink text={to} toCopy={false} />
+          </small>
+        </p>
+        <p>
+          <strong>Descriptions: </strong>{' '}
+          <small>{recommendedCallUsed.toName}</small>
+        </p>
+        <p>
+          <strong>Function: </strong>
+          <small>{recommendedCallUsed.functionName}</small>
+        </p>
+        <p>
+          <strong>Function Signature: </strong>{' '}
+          <small>{encodedFunctionName}</small>
+        </p>
+        <strong>Params: </strong>
+        {Object.keys(callParameters).map(paramIndex => {
+          return (
+            <p>
+              <small>{callParameters[paramIndex]} </small>
+            </p>
+          );
+        })}
+        <strong>data: </strong>
+        <small>{data} </small>
+      </div>
+    );
+  }
+  return (
+    <div>
+      <small>{decodedCallText}</small>
+    </div>
+  );
+};
